content/html/content/test/file_iframe_sandbox_c_if5.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 341604</title>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

</head>
<script type="text/javascript">
  function ok(result, desc) {
    window.parent.ok_wrapper(result, desc);
  }
</script>
<body onLoad="doStuff()">
  I am sandboxed but with "allow-same-origin"

  <a href = 'javascript:ok(false, "documents sandboxed without allow-scripts should not be able to run script with javascript: URLs");' id='a_link'>click me</a>
</body>
</html>
